+++ src/pages/soojin/Main/Feed.scss
@@ -1,0 +1,79 @@
+@import '../../../styles/variables.scss';
+
+.article {
+  top: 10%;
+  width: 60%;
+  background-color: white;
+  border: $border-style;
+
+  .profile {
+    @include flex-center;
+
+    img {
+      width: 50px;
+      margin: 10px;
+    }
+
+    h1 {
+      font-size: 14px;
+      @include hs;
+    }
+
+    h2 {
+      @include hs;
+      font-size: 12px;
+      font-weight: normal;
+    }
+  }
+
+  .photo {
+    img {
+      width: 100%;
+    }
+  }
+
+  .information {
+    padding-left: 5px;
+    ul {
+      @include flex-center;
+      img {
+        width: 35px;
+        margin: 2px;
+      }
+    }
+    .numberOfLikes {
+      @include flex-center;
+      font-size: 14px;
+      img {
+        width: 30px;
+      }
+    }
+    .explanation {
+      @include flex-center;
+      font-size: 14px;
+      margin: 5px;
+      .boldId {
+        font-weight: bold;
+        margin-right: 5px;
+      }
+    }
+  }
+
+  .commentWrapper {
+    display: flex;
+    font-size: 14px;
+    .yourId {
+      font-weight: bold;
+      margin: 2px 12px;
+    }
+  }
+
+  .comments {
+    input {
+      display: flex;
+      width: 100%;
+      height: 30px;
+      margin-top: 10px;
+    }
+  }
+}
